application split into seperate screens
