+++ server/routes/reportDeleted.ts
@@ -1,0 +1,25 @@
+import { Router } from 'express'
+import AuditService, { Page } from '../services/auditService'
+import BreachNoticeApiClient, { BreachNotice } from '../data/breachNoticeApiClient'
+import { HmppsAuthClient } from '../data'
+
+export default function reportDeletedRoutes(
+  router: Router,
+  auditService: AuditService,
+  hmppsAuthClient: HmppsAuthClient,
+): Router {
+  router.get('/report-deleted/:id', async (req, res, next) => {
+    await auditService.logPageView(Page.REPORT_DELETED, { who: res.locals.user.username, correlationId: req.id })
+    const breachNoticeApiClient = new BreachNoticeApiClient(
+      await hmppsAuthClient.getSystemClientToken(res.locals.user.username),
+    )
+    const breachNoticeId = req.params.id
+    let breachNotice: BreachNotice = null
+    breachNotice = await breachNoticeApiClient.getBreachNoticeById(breachNoticeId as string)
+    res.render('pages/report-deleted', {
+      breachNotice,
+    })
+  })
+
+  return router
+}
